move submessage elements tests
