+++ dds/src/implementation/rtps_udp_psm/mapping_rtps_messages/submessage_elements/count.rs
@@ -3,7 +3,7 @@ use std::io::{Error, Write};
 use byteorder::ByteOrder;
 
 use crate::implementation::{
-    rtps::{types::Count},
+    rtps::types::Count,
     rtps_udp_psm::mapping_traits::{
         MappingReadByteOrdered, MappingWriteByteOrdered, NumberOfBytes,
     },
@@ -26,7 +26,6 @@ impl<'de> MappingReadByteOrdered<'de> for Count {
     }
 }
 
-
 impl NumberOfBytes for Count {
     fn number_of_bytes(&self) -> usize {
         4
@@ -35,9 +34,8 @@ impl NumberOfBytes for Count {
 
 #[cfg(test)]
 mod tests {
-    use crate::implementation::rtps_udp_psm::mapping_traits::{from_bytes_le, to_bytes_le};
-
     use super::*;
+    use crate::implementation::rtps_udp_psm::mapping_traits::to_bytes_le;
 
     #[test]
     fn serialize_guid_prefix() {
@@ -47,18 +45,6 @@ mod tests {
             vec![
             7, 0, 0,0 , //value (long)
         ]
-        );
-    }
-
-    #[test]
-    fn deserialize_guid_prefix() {
-        let expected = Count::new(7);
-        assert_eq!(
-            expected,
-            from_bytes_le(&[
-            7, 0, 0,0 , //value (long)
-        ])
-            .unwrap()
         );
     }
 }
