+++ dds/src/implementation/rtps_udp_psm/mapping_rtps_messages/submessage_elements/fragment_number_set.rs
@@ -4,10 +4,10 @@ use byteorder::ByteOrder;
 
 use crate::implementation::{
     rtps::messages::{submessage_elements::FragmentNumberSet, types::FragmentNumber},
-    rtps_udp_psm::mapping_traits::{MappingReadByteOrdered, MappingWriteByteOrdered, NumberOfBytes},
+    rtps_udp_psm::mapping_traits::{
+        MappingReadByteOrdered, MappingWriteByteOrdered, NumberOfBytes,
+    },
 };
-
-
 
 impl NumberOfBytes for FragmentNumberSet {
     fn number_of_bytes(&self) -> usize {
@@ -15,7 +15,8 @@ impl NumberOfBytes for FragmentNumberSet {
             *max - self.base + FragmentNumber::new(1)
         } else {
             FragmentNumber::new(0)
-        }.into();
+        }
+        .into();
         let number_of_bitmap_elements = (num_bits as usize + 31) / 32; // aka "M"
         8 /*bitmapBase + numBits */ + 4 * number_of_bitmap_elements /* bitmap[0] .. bitmap[M-1] */
     }
@@ -49,29 +50,33 @@ impl MappingWriteByteOrdered for FragmentNumberSet {
 
 impl<'de> MappingReadByteOrdered<'de> for FragmentNumberSet {
     fn mapping_read_byte_ordered<B: ByteOrder>(buf: &mut &'de [u8]) -> Result<Self, Error> {
-            let base: u32 = MappingReadByteOrdered::mapping_read_byte_ordered::<B>(buf)?;
-            let num_bits: u32 = MappingReadByteOrdered::mapping_read_byte_ordered::<B>(buf)?;
-            let number_of_bitmap_elements = ((num_bits + 31) / 32) as usize; //In standard refered to as "M"
-            let mut bitmap = [0; 8];
-            for bitmap_i in bitmap.iter_mut().take(number_of_bitmap_elements) {
-                *bitmap_i = MappingReadByteOrdered::mapping_read_byte_ordered::<B>(buf)?;
-            }
+        let base: u32 = MappingReadByteOrdered::mapping_read_byte_ordered::<B>(buf)?;
+        let num_bits: u32 = MappingReadByteOrdered::mapping_read_byte_ordered::<B>(buf)?;
+        let number_of_bitmap_elements = ((num_bits + 31) / 32) as usize; //In standard refered to as "M"
+        let mut bitmap = [0; 8];
+        for bitmap_i in bitmap.iter_mut().take(number_of_bitmap_elements) {
+            *bitmap_i = MappingReadByteOrdered::mapping_read_byte_ordered::<B>(buf)?;
+        }
 
-            let mut set = Vec::with_capacity(256);
-            for delta_n in 0..num_bits as usize {
-                if (bitmap[delta_n / 32] & (1 << (31 - delta_n % 32))) == (1 << (31 - delta_n % 32)) {
-                    set.push(FragmentNumber::new(base + delta_n as u32));
-                }
+        let mut set = Vec::with_capacity(256);
+        for delta_n in 0..num_bits as usize {
+            if (bitmap[delta_n / 32] & (1 << (31 - delta_n % 32))) == (1 << (31 - delta_n % 32)) {
+                set.push(FragmentNumber::new(base + delta_n as u32));
             }
-            Ok(Self { base: FragmentNumber::new(base), set})
+        }
+        Ok(Self {
+            base: FragmentNumber::new(base),
+            set,
+        })
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::implementation::{rtps_udp_psm::mapping_traits::{from_bytes_le, to_bytes_le}, rtps::messages::types::FragmentNumber};
-
     use super::*;
+    use crate::implementation::{
+        rtps::messages::types::FragmentNumber, rtps_udp_psm::mapping_traits::to_bytes_le,
+    };
 
     #[test]
     fn serialize_fragment_number_max_gap() {
@@ -92,28 +97,5 @@ mod tests {
             0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[6] (long)
             0b000_0001, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[7] (long)
         ]);
-    }
-
-    #[test]
-    fn deserialize_fragment_number_set_max_gap() {
-        let expected = FragmentNumberSet {
-            base: FragmentNumber::new(2),
-            set: vec![FragmentNumber::new(2), FragmentNumber::new(257)],
-        };
-        #[rustfmt::skip]
-        let result = from_bytes_le(&[
-            2, 0, 0, 0, // bitmapBase: (unsigned long)
-            0, 1, 0, 0, // numBits (unsigned long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_1000_0000, // bitmap[0] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[1] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[2] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[3] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[4] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[5] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[6] (long)
-            0b000_0001, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[7] (long)
-
-        ]).unwrap();
-        assert_eq!(expected, result);
     }
 }
