+++ dds/src/implementation/rtps_udp_psm/mapping_rtps_messages/submessage_elements/parameter_list.rs
@@ -107,10 +107,8 @@ impl<'de> MappingReadByteOrdered<'de> for ParameterList {
 
 #[cfg(test)]
 mod tests {
-
-    use crate::implementation::rtps_udp_psm::mapping_traits::{from_bytes_le, to_bytes_le};
-
     use super::*;
+    use crate::implementation::rtps_udp_psm::mapping_traits::{to_bytes_le};
 
     #[test]
     fn serialize_parameter() {
@@ -147,30 +145,6 @@ mod tests {
     }
 
     #[test]
-    fn deserialize_parameter_non_multiple_of_4() {
-        let expected = Parameter::new(ParameterId(2), vec![5, 6, 7, 8, 9, 10, 11, 0]);
-        #[rustfmt::skip]
-        let result = from_bytes_le(&[
-            0x02, 0x00, 8, 0, // Parameter | length
-            5, 6, 7, 8,       // value
-            9, 10, 11, 0,     // value
-        ]).unwrap();
-        assert_eq!(expected, result);
-    }
-
-    #[test]
-    fn deserialize_parameter() {
-        let expected = Parameter::new(ParameterId(2), vec![5, 6, 7, 8, 9, 10, 11, 12]);
-        #[rustfmt::skip]
-        let result = from_bytes_le(&[
-            0x02, 0x00, 8, 0, // Parameter | length
-            5, 6, 7, 8,       // value
-            9, 10, 11, 12,       // value
-        ]).unwrap();
-        assert_eq!(expected, result);
-    }
-
-    #[test]
     fn serialize_parameter_list() {
         let parameter_1 = Parameter::new(ParameterId(2), vec![51, 61, 71, 81]);
         let parameter_2 = Parameter::new(ParameterId(3), vec![52, 62, 0, 0]);
@@ -196,96 +170,4 @@ mod tests {
         assert_eq!(parameter.number_of_bytes(), 4);
     }
 
-    #[test]
-    fn deserialize_parameter_list() {
-        let expected = ParameterList::new(vec![
-            Parameter::new(ParameterId(2), vec![15, 16, 17, 18]),
-            Parameter::new(ParameterId(3), vec![25, 26, 27, 28]),
-        ]);
-        #[rustfmt::skip]
-        let result = from_bytes_le(&[
-            0x02, 0x00, 4, 0, // Parameter ID | length
-            15, 16, 17, 18,        // value
-            0x03, 0x00, 4, 0, // Parameter ID | length
-            25, 26, 27, 28,        // value
-            0x01, 0x00, 0, 0, // Sentinel: Parameter ID | length
-        ]).unwrap();
-        assert_eq!(expected, result);
-    }
-
-    #[test]
-    fn deserialize_parameter_list_with_long_parameter_including_sentinel() {
-        #[rustfmt::skip]
-        let parameter_value_expected = vec![
-            0x01, 0x00, 0x00, 0x00,
-            0x01, 0x00, 0x00, 0x00,
-            0x01, 0x01, 0x01, 0x01,
-            0x01, 0x01, 0x01, 0x01,
-            0x01, 0x01, 0x01, 0x01,
-            0x01, 0x01, 0x01, 0x01,
-        ];
-
-        let expected = ParameterList::new(vec![Parameter::new(
-            ParameterId(0x32),
-            parameter_value_expected,
-        )]);
-        #[rustfmt::skip]
-        let result = from_bytes_le(&[
-            0x32, 0x00, 24, 0x00, // Parameter ID | length
-            0x01, 0x00, 0x00, 0x00, // Parameter value
-            0x01, 0x00, 0x00, 0x00, // Parameter value
-            0x01, 0x01, 0x01, 0x01, // Parameter value
-            0x01, 0x01, 0x01, 0x01, // Parameter value
-            0x01, 0x01, 0x01, 0x01, // Parameter value
-            0x01, 0x01, 0x01, 0x01, // Parameter value
-            0x01, 0x00, 0x00, 0x00, // PID_SENTINEL, Length: 0
-        ]).unwrap();
-        assert_eq!(expected, result);
-    }
-
-    #[test]
-    fn deserialize_parameter_list_with_multiple_parameters_with_same_id() {
-        #[rustfmt::skip]
-        let parameter_value_expected1 = vec![
-            0x01, 0x00, 0x00, 0x00,
-            0x01, 0x00, 0x00, 0x00,
-            0x01, 0x01, 0x01, 0x01,
-            0x01, 0x01, 0x01, 0x01,
-            0x01, 0x01, 0x01, 0x01,
-            0x01, 0x01, 0x01, 0x01,
-        ];
-        #[rustfmt::skip]
-        let parameter_value_expected2 = vec![
-            0x01, 0x00, 0x00, 0x00,
-            0x01, 0x00, 0x00, 0x00,
-            0x02, 0x02, 0x02, 0x02,
-            0x02, 0x02, 0x02, 0x02,
-            0x02, 0x02, 0x02, 0x02,
-            0x02, 0x02, 0x02, 0x02,
-        ];
-
-        let expected = ParameterList::new(vec![
-            Parameter::new(ParameterId(0x32), parameter_value_expected1),
-            Parameter::new(ParameterId(0x32), parameter_value_expected2),
-        ]);
-        #[rustfmt::skip]
-        let result = from_bytes_le(&[
-            0x32, 0x00, 24, 0x00, // Parameter ID | length
-            0x01, 0x00, 0x00, 0x00, // Parameter value
-            0x01, 0x00, 0x00, 0x00, // Parameter value
-            0x01, 0x01, 0x01, 0x01, // Parameter value
-            0x01, 0x01, 0x01, 0x01, // Parameter value
-            0x01, 0x01, 0x01, 0x01, // Parameter value
-            0x01, 0x01, 0x01, 0x01, // Parameter value
-            0x32, 0x00, 24, 0x00, // Parameter ID | length
-            0x01, 0x00, 0x00, 0x00, // Parameter value
-            0x01, 0x00, 0x00, 0x00, // Parameter value
-            0x02, 0x02, 0x02, 0x02, // Parameter value
-            0x02, 0x02, 0x02, 0x02, // Parameter value
-            0x02, 0x02, 0x02, 0x02, // Parameter value
-            0x02, 0x02, 0x02, 0x02, // Parameter value
-            0x01, 0x00, 0x00, 0x00, // PID_SENTINEL, Length: 0
-        ]).unwrap();
-        assert_eq!(expected, result);
-    }
 }
