+++ dds/src/implementation/rtps_udp_psm/mapping_rtps_messages/submessage_elements/sequence_number_set.rs
@@ -81,9 +81,8 @@ impl<'de> MappingReadByteOrdered<'de> for SequenceNumberSet {
 
 #[cfg(test)]
 mod tests {
-    use crate::implementation::rtps_udp_psm::mapping_traits::{from_bytes_le, to_bytes_le};
-
     use super::*;
+    use crate::implementation::rtps_udp_psm::mapping_traits::to_bytes_le;
 
     #[test]
     fn serialize_sequence_number_max_gap() {
@@ -105,29 +104,6 @@ mod tests {
             0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[6] (long)
             0b000_0001, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[7] (long)
         ]);
-    }
-
-    #[test]
-    fn deserialize_sequence_number_set_max_gap() {
-        let expected = SequenceNumberSet {
-            base: SequenceNumber::new(2),
-            set: vec![SequenceNumber::new(2), SequenceNumber::new(257)],
-        };
-        #[rustfmt::skip]
-        let result = from_bytes_le(&[
-            0, 0, 0, 0, // bitmapBase: high (long)
-            2, 0, 0, 0, // bitmapBase: low (unsigned long)
-            0, 1, 0, 0, // numBits (unsigned long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_1000_0000, // bitmap[0] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[1] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[2] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[3] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[4] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[5] (long)
-            0b000_0000, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[6] (long)
-            0b000_0001, 0b_0000_0000, 0b_0000_0000, 0b_0000_0000, // bitmap[7] (long)
-        ]).unwrap();
-        assert_eq!(expected, result);
     }
 
     #[test]
