+++ dds/src/implementation/rtps_udp_psm/mapping_rtps_messages/submessage_elements/vendor_id.rs
@@ -26,20 +26,11 @@ impl<'de> MappingReadByteOrdered<'de> for VendorId {
 
 #[cfg(test)]
 mod tests {
-    use crate::implementation::{
-        rtps::types::VendorId,
-        rtps_udp_psm::mapping_traits::{from_bytes_le, to_bytes_le},
-    };
+    use crate::implementation::{rtps::types::VendorId, rtps_udp_psm::mapping_traits::to_bytes_le};
 
     #[test]
     fn serialize_vendor_id() {
         let data = VendorId::new([1, 2]);
         assert_eq!(to_bytes_le(&data).unwrap(), vec![1, 2,]);
-    }
-
-    #[test]
-    fn deserialize_vendor_id() {
-        let expected = VendorId::new([1, 2]);
-        assert_eq!(expected, from_bytes_le(&[1, 2,]).unwrap());
     }
 }
